화장실 조회 데이터 저장 성공, 로그아웃쪽 경로 다시 수정

+++ src/main/java/global/sesoc/brr/vo/DayAverVO.java
@@ -10,8 +10,8 @@ import lombok.NoArgsConstructor;
 public class DayAverVO {
 
 	
-	private int lat;
-	private int lng;
+	private double lat;
+	private double lng;
 	private String toiletnm;
 	private String id;
 	private String inputdate;
